refactor(demo): use umi as project cli
