+++ demo/package.json
@@ -5,35 +5,37 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack serve --port 8881"
+    "start": "umi dev",
+    "postinstall": "umi generate tmp",
+    "prettier": "prettier --write '**/*.{js,jsx,tsx,ts,less,md,json}'",
+    "test": "umi-test",
+    "test:coverage": "umi-test --coverage"
   },
   "author": {
     "name": "AntV",
     "url": "https://antv.vision/"
   },
   "license": "MIT",
+  "gitHooks": {
+    "pre-commit": "lint-staged"
+  },
+  "lint-staged": {
+    "*.{js,jsx,less,md,json}": [
+      "prettier --write"
+    ],
+    "*.ts?(x)": [
+      "prettier --parser=typescript --write"
+    ]
+  },
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.14.7",
-    "@babel/plugin-transform-runtime": "^7.16.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@types/lodash": "^4.14.165",
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "babel-loader": "^8.2.2",
-    "babel-plugin-import": "^1.13.3",
-    "crypto-browserify": "^3.12.0",
-    "css-loader": "^6.2.0",
-    "file-loader": "^6.2.0",
-    "less": "^4.1.1",
-    "less-loader": "^10.0.1",
-    "style-loader": "^3.2.1",
-    "webpack": "^5.48.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@umijs/preset-react": "1.x",
+    "@umijs/test": "^3.5.20",
+    "lint-staged": "^10.0.7",
+    "prettier": "^2.2.0",
+    "typescript": "^4.1.2",
+    "yorkie": "^2.0.0"
   },
   "dependencies": {
     "@ant-design/icons": "^4.6.3",
@@ -59,6 +61,7 @@
     "react-dom": "^17.0.2",
     "react-json-view": "^1.19.1",
     "react-vega": "^7.4.3",
+    "umi": "^3.5.20",
     "vega": "^5.15.0",
     "vega-datasets": "^2.2.0",
     "vega-lite": "^4.15.0"
