--- demo/src/index.html
@@ -1,13 +1,0 @@
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Test</title>
-    <script src="https://gw.alipayobjects.com/os/lib/react/16.10.2/umd/react.development.js"></script>
-    <script src="https://gw.alipayobjects.com/os/lib/react-dom/16.10.2/umd/react-dom.development.js"></script>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script src="/bundle.js"></script>
-  </body>
-</html>
