--- demo/src/index.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import App from './app';
-import './index.less';
-
-ReactDOM.render(<App />, document.querySelector('#root'));
