+++ demo/src/pages/index.less
@@ -1,21 +1,24 @@
-@import 'antd/dist/antd.less';
+@import '~@antv/auto-chart/esm/index.css';
 
 .chart-container {
   height: 600px;
   display: flex;
   flex-direction: column;
   margin-bottom: 32px;
+
   &-title {
     flex: none;
     font-size: 28px;
     margin-bottom: 16px;
   }
+
   &-description {
     flex: none;
     color: #999;
     font-size: 13px;
     margin-bottom: 16px;
   }
+
   .canvas-container {
     flex: 1;
   }
@@ -67,7 +70,7 @@ nav {
   border-right: 1px solid #f0f0f0;
 }
 
-.ant-layout.ant-layout-has-sider > .ant-layout-content {
+.ant-layout.ant-layout-has-sider>.ant-layout-content {
   overflow-x: scroll;
 }
