--- demo/webpack.config.js
@@ -1,85 +1,0 @@
-const path = require('path');
-const webpack = require('webpack');
-
-const devConfig = {
-  mode: 'development',
-  devtool: 'cheap-module-source-map',
-  entry: {
-    app: './src/index',
-  },
-  output: {
-    path: path.join(__dirname, 'public'),
-    filename: 'bundle.js',
-    chunkFilename: '[chunkhash].js',
-    publicPath: '/',
-  },
-  externals: {
-    react: 'React',
-    'react-dom': 'ReactDOM',
-  },
-  target: 'web',
-  resolve: {
-    extensions: ['.js', '.ts', '.tsx', '.json'],
-    fallback: { path: false, buffer: false, stream: false, crypto: require.resolve('crypto-browserify'), fs: false },
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(t|j)sx?$/,
-        exclude: /node_modules/,
-        loader: 'babel-loader',
-        options: {
-          presets: ['@babel/preset-typescript', ['@babel/preset-env', { modules: false }], '@babel/preset-react'],
-          plugins: [
-            '@babel/plugin-transform-runtime',
-            '@babel/plugin-proposal-class-properties',
-            '@babel/plugin-proposal-object-rest-spread',
-          ],
-        },
-      },
-      {
-        test: /\.(css|less)$/,
-        use: [
-          { loader: 'style-loader' },
-          { loader: 'css-loader', options: { sourceMap: true } },
-          {
-            loader: 'less-loader',
-            options: {
-              sourceMap: true,
-              lessOptions: {
-                javascriptEnabled: true,
-              },
-            },
-          },
-        ],
-      },
-      {
-        test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          {
-            loader: 'file-loader',
-            options: {
-              name: '[name].[ext]',
-              outputPath: 'fonts/',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  devServer: {
-    disableHostCheck: true,
-    host: '0.0.0.0',
-    hot: true,
-    contentBase: path.join(__dirname, 'src'),
-    publicPath: '/',
-  },
-  plugins: [
-    new webpack.IgnorePlugin({ resourceRegExp: /^(fs|child_process)$/ }),
-    new webpack.EnvironmentPlugin({
-      NODE_ENV: 'development',
-    }),
-  ],
-};
-
-module.exports = devConfig;
